feat(phase-aj): carry caller activity on CLI requests
